rename PlantUML grammar
