+++ besser/BUML/metamodel/ocl/rules.py
